lab3/requirements.txt
sqlmodel>=0.0.14
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0

